Get version number from git
Not exceptionally useful for stable snaps which build from a know version,
but very practical for building from master.

snapcraft can also do this automatically by setting version to 'git', but
the numbers it generates are too generic and don't look nice (e.g.
'0+git.badcoffee-dirty'). We know the tagging scheme used by Liferea, so
we can do this from a scriptlet and get 1.13.8 when building from a commit
tagged with v1.13.8 or 1.13.8-10-600dcoffee when building from a later one.